requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
google-genai
opik
passlib[bcrypt]
pyjwt
python-multipart
email-validator
